update dated 22NOv

+++ sv-api/src/main/java/com/sv/app/repository/AttachedDocumentRepository.java
@@ -2,10 +2,9 @@ package com.sv.app.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.sv.app.bean.ApplicationStatusBean;
 import com.sv.app.bean.AttachedDocumentBean;
 
-public interface AttachedDocumentRepository extends CrudRepository<AttachedDocumentBean, Long>{
+public interface AttachedDocumentRepository extends CrudRepository<AttachedDocumentBean, Integer>{
 
 
 }
